blas: implement blas routines
